Deploy Changes

+++ app/Http/Livewire/PostComponent.php
@@ -12,7 +12,7 @@ use Illuminate\Support\Facades\Auth;
 use Carbon\Carbon;
 
 class PostComponent extends Component
-{  
+{
     protected $listeners=['postsUpdated'=>'$refresh'];
 
     public $search="",$city="all",$category="all";
@@ -30,7 +30,7 @@ class PostComponent extends Component
         }
         else if($this->filter==$cid){
             $posts=Post::where('company_id',$this->filter)->get();
-        }   
+        }
         else if($this->presents=="presents"){
             $posts=Auth::user()->presents;
         }
@@ -45,14 +45,14 @@ class PostComponent extends Component
                 else if($this->city!="all" && $this->category=="all")
                     $posts=Post::where('description','like','%'.$this->search.'%')->where('city',$this->city)->orderByDesc('id')->get();
                 else
-                    $posts=Post::where('description','like','%'.$this->search.'%')->where('category',$this->category)->orderByDesc('id')->get(); 
+                    $posts=Post::where('description','like','%'.$this->search.'%')->where('category',$this->category)->orderByDesc('id')->get();
             }
             else{
                 $posts=Post::orderByDesc('id')->get();
             }
         }
         return view('livewire.post-component',compact('posts'));
-    
+
     }
     public function setPostToDelete($postDelete){
         $this->postToDelete=$postDelete;
@@ -63,7 +63,8 @@ class PostComponent extends Component
     }
     public function save($postId)
     {
-       
+        if(!Auth::check())
+            return $this->redirect('/login');
         if(Auth::user()->isSaved($postId)){
             DB::table('post_user')->where('user_id',Auth::id())->where('post_id',$postId)->delete();
         }else{
@@ -78,6 +79,8 @@ class PostComponent extends Component
     }
 
     public function presentJob($postId){
+        if(!Auth::check())
+            return $this->redirect('/login');
         if(Auth::user()->isPresented($postId)){
             DB::table('job_user')->where('user_id',Auth::id())->where('post_id',$postId)->delete();
             DB::table('notifications')->where('user_id',Auth::id())->where('post_id',$postId)->delete();
@@ -97,12 +100,14 @@ class PostComponent extends Component
                 'created_at'=>Carbon::now()->toDateTimeString(),
             ]);
         }
-        
+
         $this->emit('postUpdated');
         $this->emitTo('notification-counter','newNotification');
     }
 
     public function like($pid){
+        if(!Auth::check())
+            return $this->redirect('/login');
         if(Auth::user()->isLiked($pid)){
             DB::table('post_like')->where('user_id',Auth::id())->where('post_id',$pid)->delete();
         }else{
@@ -120,6 +125,8 @@ class PostComponent extends Component
         $this->postId=$post_id;
     }
     public function addComment(){
+        if(!Auth::check())
+            return $this->redirect('/login');
         Comment::create([
             'user_id'=>Auth::id(),
             'post_id'=>$this->postId,
@@ -139,6 +146,8 @@ class PostComponent extends Component
         }
     }
     public function addReply($commentId){
+        if(!Auth::check())
+            return $this->redirect('/login');
         Reply::create([
             'user_id'=>Auth::id(),
             'comment_id'=>$commentId,
@@ -160,5 +169,10 @@ class PostComponent extends Component
     public function resetReplies(){
         $this->allReplies=Comment::find($this->reply)->replies;
     }
-    
+    public function replyComment($id)
+    {
+        if(!Auth::check())
+            return $this->redirect('/login');
+    }
+
 }
